+++ src/algorithms/search/ComperBestFirstSearch.java
@@ -9,7 +9,6 @@ public class ComperBestFirstSearch implements Comparator<AState>{
      */
     @Override
     public int compare(AState o1, AState o2) {
-
-        return Integer.compare(o1.getCost(),o2.getCost());
+        return Double.compare(o1.getCost(),o2.getCost());
     }
 }
